Stub out disconnect page in resources.
In WSGI, a close method on the iterable returned by a WSGI callable MUST be
called after the response is sent. So it shouldn't be too difficult to fire
that for JSON and template resources. However, for socket resources, we don't
have a one to one with HTTP requests; we only want to run the disconnect page
when the *socket* closes. So that's going to take some more thought.