Refactor aspen.tests into aspen.testing and tests/
Aspen should bundle helpers for testing Simplates. This commit is a step in
that direction. Now aspen.testing contains helpers and fixture and such, and
the new tests/ directory in the root contains Aspen's own test suite, which
uses the aspen.testing module. It turns out that the existing helpers aren't
that helpful for testing actual websites, so I'm not adding any documentation
as part of this commit.