Implement indirect negotiation
The canonical way to do content negotiation is via the Accept header.
But of course that's rather inconvenient to set compared to modifying
the URL. This implements "indirect negotiation", so that /foo.html and
/foo.txt both resolve to /foo (if that exists), and the content
negotiation is performed on the basis of a media type inferred from the
file extension in the URL. Good suggestion, @pjz. Fixes #56.