Fix regression where virtual paths were overeager
If there's a virtual path directory in the root of a site, then it will
swallow requests for /favicon.ico and /robots.txt if those files are not
otherwise present. We don't want that. For /robots.txt we want a 404,
and for /favicon.ico we actually want to serve a default bundled Aspen
one. Fixes #48