Removes code duplication in aspen.utils.to_rfc822
The logic used in to_rfc822 also exists in the standard library
module email.utils. As RFC 2822 is an email specification, this
seems to be the logical place to centralize the logic. Further,
email.utils has no dependencies outside the standard library.

I've imported email.utils as email_utils to avoid possible
confusion between the name of this module (utils) and the imported
module, without incurring the overhead of importing the entire
email module.