Updates tests to better comply with RFC 822
RFC 0822 (http://www.ietf.org/rfc/rfc0822.txt), page 26, requires
that a numerically-defined GMT offset begin with a - or +, but does
not specify which much be used for an offset of zero. As such, I've
updated the tests to allow either sign to be used where we test for
the presence of zero offset.