Avoid re-raising Responses from simplates; #153
I ended up factoring out the code in Website.log_access to pull the
filename/linenum info out of the current Response exception and putting
that on the response object itself. This makes testing easier and is
nice encapsulation. This also gave me an opportunity to `del tb`, which
we needed to do to avoid the memory leak associated with circular refs
from sys.exc_info. Includes a couple tests.