Fix last socket test; #148
This test was originally breaking because we weren't properly starting a
socket thread because the filename was wrong and the content was wrong.
Now that the test is passing we needed to bring back the part where we
*close* the socket thread. That's where the slowdown was from. It's nice
though: the mimetypes.init thing still gives us almost a 4x boost in
test time. !m @pjz