Add test client helpers for catching Response
If you set raise_immediately=False when testing, then who knows what
you're going to get? When a test raises an exception you want exactly
that exception, not a downstream wrapped who knows what. So instead of
recommending overriding raise_immediately per-test, here are a couple
extra helpers: GxT and PxST, which are just like GET and POST but they
expect to have an exception raised, and they return that.