Only support negotiated simplates for errors
This is a step on the way to addressing #279. It makes error handling a
lot simpler if we don't look for {$code,error}{.$ext,.$ext.spt,.spt},
where $ext is infinitely variable. Instead we look just for
{$code,error}.spt. Eventually we may decide to simplify non-error
simplates the same way: only .spt, no special-casing of .$ext.spt on the
filesystem (#286). For now though the next step for #279 would be to
update the Accept header of the internal redirect in
delegate_error_to_simplate to enforce the media type that had been
negotiated to begin with.