Fix the rest of test_website.py
This subtly alters the behavior of error simplates wrt content
negotiation. Now we look at dispatch_result.extra['accept'] (i.e., the
URL path) instead of the resource.media_type (i.e., the filesystem path)
when deciding how to clobber the Accept header that we pass on to the
error simplate. This machinery could probably get a refactoring of it's
own, but I didn't want to creep scope here.